+++ common/cosmwasm-smart-contracts/service-provider-directory/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod events;
 pub mod msg;
+pub mod response;
 pub mod types;
 
 // Re-export all types at the top-level
